+++ .github/workflows/pr-redspace-preview.yml
@@ -1,13 +1,17 @@
 name: PR Redspace Preview
 
 on:
-  # DS14-start: run only after the PR RSI validator workflow completes.
-  workflow_run:
-    workflows:
-      - RSI Validator
+  pull_request_target:
     types:
-      - completed
-  # DS14-end
+      - opened
+      - reopened
+      - synchronize
+      - ready_for_review
+    paths:
+      - 'Resources/Prototypes/**'
+      - 'Resources/Textures/**'
+      - 'Tools/pr_redspace_preview.py'
+      - '.github/workflows/pr-redspace-preview.yml'
   workflow_dispatch:
 
 permissions:
@@ -18,55 +22,17 @@ jobs:
   preview:
     name: Build Preview
     runs-on: ubuntu-latest
-    # DS14-start: only build for pull_request RSI validator runs or manual dispatches.
-    if: >-
-      github.event_name == 'workflow_dispatch' ||
-      (
-        github.event.workflow_run.event == 'pull_request' &&
-        github.event.workflow_run.name == 'RSI Validator'
-      )
-    # DS14-end
+    if: github.event_name != 'pull_request_target' || github.event.pull_request.draft == false
     steps:
       - name: Checkout base repository
         uses: actions/checkout@v4.2.2
         with:
           fetch-depth: 0
 
-      - name: Resolve PR context
-        # DS14-start: recover PR metadata from the finished RSI validator run.
-        if: github.event_name == 'workflow_run'
-        id: pr_context
-        env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          REPOSITORY: ${{ github.repository }}
-          PR_NUMBER: ${{ github.event.workflow_run.pull_requests[0].number }}
-        run: |
-          gh api "repos/${REPOSITORY}/pulls/${PR_NUMBER}" > pr.json
-          echo "number=${PR_NUMBER}" >> "$GITHUB_OUTPUT"
-          echo "base_sha=$(jq -r '.base.sha' pr.json)" >> "$GITHUB_OUTPUT"
-          echo "head_sha=$(jq -r '.head.sha' pr.json)" >> "$GITHUB_OUTPUT"
-          echo "head_repo=$(jq -r '.head.repo.clone_url' pr.json)" >> "$GITHUB_OUTPUT"
-        # DS14-end
-
       - name: Fetch PR head commit
-        # DS14: workflow_run provides the validated PR head repository and SHA.
-        if: github.event_name == 'workflow_run'
-        env:
-          HEAD_REPO: ${{ steps.pr_context.outputs.head_repo }}
-          HEAD_SHA: ${{ steps.pr_context.outputs.head_sha }}
+        if: github.event_name == 'pull_request_target'
         run: |
-          git fetch --no-tags "$HEAD_REPO" "$HEAD_SHA"
-
-      - name: Resolve comparison base
-        id: compare_base
-        env:
-          # DS14: workflow_run uses the PR that triggered RSI validation.
-          BASE_SHA: ${{ github.event_name == 'workflow_run' && steps.pr_context.outputs.base_sha || github.sha }}
-          HEAD_SHA: ${{ github.event_name == 'workflow_run' && steps.pr_context.outputs.head_sha || github.sha }}
-        run: |
-          COMPARE_BASE_SHA="$(git merge-base "$BASE_SHA" "$HEAD_SHA")"
-          echo "base_sha=$COMPARE_BASE_SHA" >> "$GITHUB_OUTPUT"
-          echo "head_sha=$HEAD_SHA" >> "$GITHUB_OUTPUT"
+          git fetch --no-tags --depth=1 "${{ github.event.pull_request.head.repo.clone_url }}" "${{ github.event.pull_request.head.sha }}"
 
       - name: Setup Python
         uses: actions/setup-python@v5
@@ -102,12 +68,12 @@ jobs:
             .artifacts/pr-redspace-preview.json
 
       - name: Publish preview image to preview branch
-        if: github.event_name == 'workflow_run'
+        if: github.event_name == 'pull_request_target'
         id: publish
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           REPOSITORY: ${{ github.repository }}
-          PR_NUMBER: ${{ steps.pr_context.outputs.number }}
+          PR_NUMBER: ${{ github.event.pull_request.number }}
           HEAD_SHA: ${{ steps.render.outputs.head_sha }}
         run: |
           set -euo pipefail
@@ -144,19 +110,19 @@ jobs:
           echo "image_url=$RAW_URL" >> "$GITHUB_OUTPUT"
 
       - name: Find existing preview comment
-        if: github.event_name == 'workflow_run'
+        if: github.event_name == 'pull_request_target'
         id: find_comment
         uses: peter-evans/find-comment@v3
         with:
-          issue-number: ${{ steps.pr_context.outputs.number }}
+          issue-number: ${{ github.event.pull_request.number }}
           comment-author: 'github-actions[bot]'
           body-includes: '<!-- pr-redspace-preview-comment -->'
 
       - name: Create or update preview comment
-        if: github.event_name == 'workflow_run'
+        if: github.event_name == 'pull_request_target'
         uses: peter-evans/create-or-update-comment@v4
         with:
-          issue-number: ${{ steps.pr_context.outputs.number }}
+          issue-number: ${{ github.event.pull_request.number }}
           comment-id: ${{ steps.find_comment.outputs.comment-id }}
           edit-mode: replace
           body: |
